requirements.txt
asgiref==3.7.2
Django==4.2.4
djangorestframework==3.14.0
pip==23.2.1
setuptools==68.1.2
tzdata==2023.3
sentry-sdk==1.29.2
psycopg2==2.9.7
djangorestframework-jwt~=1.11.0
openai~=0.27.8